python_cli/pyproject.toml
[project]
name = "python-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Marcello Steiner", email = "steiner.marcello@gmeil.com" }]
requires-python = ">=3.12"
dependencies = [
    "absl-py>=2.1.0",
    "beautifulsoup4>=4.13.3",
    "google-adk>=1.0.0",
    "injector>=0.22.0",
    "json5>=0.12.0",
    "langchain>=0.3.21",
    "langchain-google-genai>=2.1.0",
    "loguru>=0.7.3",
    "neo4j>=5.28.1",
    "pydantic>=2.10.6",
    "random-user-agent>=1.0.1",
    "requests>=2.32.3",
    "result>=0.17.0",
    "rich>=13.9.4",
    "sqlalchemy[asyncio,mypy,postgresql-psycopg]>=2.0.41",
    "tembo-pgmq-python[async]>=0.10.0",
    "types-requests>=2.32.0.20250306",
]

[project.scripts]
python-cli = "python_cli:main._main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "basedpyright>=1.28.1",
    "mypy>=1.15.0",
    "ruff>=0.9.7",
    "types-requests>=2.32.0.20250306",
]


[tool.mypy]
disallow_untyped_defs = true
exclude = ["test/"]
ignore_missing_imports = true
no_implicit_optional = true
python_version = "3.12"
show_error_codes = true
warn_return_any = true
warn_unused_ignores = true

# RUFF section
[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "data",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "**/dist",
    "test_data",
]
# Lint and format also notebooks.
extend-include = ["**/*.ipynb"]
# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.11
target-version = "py311"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = true

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    "W",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # pydocstyle
    "D",
    # mccabe
    "C90",
    # pep8 naming
    "N",
    # flake8 annotations
    "ANN",
    #flake8 bandit
    "S",
    # blind except
    "BLE",
    # flake8 boolean trap
    "FBT",
    # flake8 builtins
    "A",
    # flake8 commas
    "COM",
    # flake8 comprehensions
    "C4",
    # flake8-datetimez (DTZ)
    "DTZ",
    # flake8-implicit-str-concat (ISC)
    "ISC",
    # flake8-logging-format (G)
    "G",
    # flake8-no-pep420 (INP)
    "INP",
    # flake8-pie
    "PIE",
    # flake8-print
    "T20",
    # flake8-pyi
    "PYI",
    # flake8-pytest-style
    "PT",
    # flake8-quotes
    "Q",
    # flake8-raise
    "R",
    # flake8-return (RET)
    "RET",
    # flake8-self
    "SLF",
    # flake8-tidy-imports
    "TID",
    # flake8-type-checking
    "TCH",
    # flake8-gettext
    "INT",
    # flake8-unused-arguments
    "ARG",
    # flake8-use-pathlib
    "PTH",
    # flake8-todos
    "TD",
    # flake8-fixme
    "FIX",
    # eradicate
    "ERA",
    # pandas-vet
    "PD",
    # Pylint
    "PL",
    # tryceratops
    "TRY",
    # flynt
    "FLY",
    # NumPy-specific rules
    "NPY",
    # Airflow
    "AIR",
    # Perflint
    "PERF",
    # refurb
    "FURB",
    # flake8-logging
    "LOG",
    # Ruff-specific rules
    "RUF",
]
ignore = ["COM812", "ISC001"]
# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

exclude = ["test/**"]

[tool.ruff.lint.pydocstyle]
convention = "google"


[tool.pyright]
include = ["src"]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    "src/experimental",
    "src/typestubs",
    ".venv/",
]
ignore = ["src/oldstuff"]

reportMissingImports = true
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportArgumentType = false
reportAny = false
pythonVersion = "3.12"
